+++ shared/src/model/config/api_user.rs
@@ -80,5 +80,3 @@ impl ProxyUserCredentialsDto {
         true
     }
 }
-
-