+++ shared/src/utils/default_utils.rs
@@ -217,6 +217,14 @@ pub fn is_default_processing_order(p: &ProcessingOrder) -> bool {
     *p == ProcessingOrder::default()
 }
 
+pub const fn default_resolve_livetv_interval() -> u32 {
+    120
+}
+
+pub const fn is_default_resolve_livetv_interval(v: &u32) -> bool {
+    *v == default_resolve_livetv_interval()
+}
+
 //////////////////////////////
 // HDHomerun Device Defaults
 //////////////////////////////